[Factory Plugin] Keep the host bootstrap in app source; no generated bridge files (br-factory-plugin-lqvd.6)
The composition test imported a generated scripts/factory-host.mts.js that
is not tracked, so CI typecheck failed. The bootstrap now lives in
src/server/factoryHost.ts and the script is a re-export shim.
