+++ apps/factory-playground/scripts/factory-host.mts.d.ts
@@ -1,0 +1,1 @@
+export * from './factory-host.mts'
