+++ apps/factory-playground/scripts/factory-host.mts.js
@@ -1,0 +1,1 @@
+export * from './factory-host.mts'
